Refactor to a new component dir, parse line breaks and fix data form
- Line breaks in the alert description should now appear as intended.
- The sampler data form should work well now and includes an additional alert
  to demonstrate the component with multiple active alerts.
- Alerts.js and the stylesheet are moved into a new component dir.
