+++ src/App.js
@@ -5,7 +5,7 @@ import logo from './mlh-prep.png';
 import WeatherMap from './components/WeatherMap/WeatherMap';
 import WeatherAlerts from './components/WeatherAlerts/WeatherAlerts';
 import cities from './assets/data/cities.json';
-import Alert from './components/Alert';
+import Alert from './components/CriticalAlerts/Alert';
 import CurrentStatus from './components/CurrentStatus';
 
 import WeatherSounds from './components/WeatherSounds/WeatherSounds';
